changed close add new text
