--- src/component/NewText.jsx
@@ -1,39 +1,0 @@
-import { useState } from "react";
-import { BackendAPI } from "../api/BackendAPIHandler";
-
-function NewText() {
-  const [text, setText] = useState("");
-
-  const handleText = (event) => setText(event.target.value);
-
-  const backendAPIInstance = new BackendAPI();
-
-  const handleAddText = async (e) => {
-    e.preventDefault();
-
-    const currentDay = new Date().setHours(0, 0, 0, 0);
-
-    try {
-      await backendAPIInstance.saveText(text, currentDay);
-      setText("");
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  return (
-    <div>
-      <h2>Add your text</h2>
-      <div className="new-text-container">
-        <form onSubmit={handleAddText}>
-          <input type="text" name="text" value={text} onChange={handleText} />
-          <button className="button" type="submit">
-            Add!
-          </button>
-        </form>
-      </div>
-    </div>
-  );
-}
-
-export default NewText;
